Rename MarketTabsModule to MarketModule
