+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/MarketFragment.kt
@@ -7,16 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseWithSearchFragment
-import io.horizontalsystems.bankwallet.modules.market.tabs.MarketTabsModule
-import io.horizontalsystems.bankwallet.modules.market.tabs.MarketTabsService
-import io.horizontalsystems.bankwallet.modules.market.tabs.MarketTabsViewModel
 import io.horizontalsystems.bankwallet.modules.transactions.FilterAdapter
 import io.horizontalsystems.core.navGraphViewModels
 import kotlinx.android.synthetic.main.fragment_market.*
 
 class MarketFragment : BaseWithSearchFragment(), FilterAdapter.Listener {
     private val filterAdapter = FilterAdapter(this)
-    private val marketTabsViewModel by viewModels<MarketTabsViewModel> { MarketTabsModule.Factory() }
+    private val marketViewModel by viewModels<MarketViewModel> { MarketModule.Factory() }
     private val navigationViewModel by navGraphViewModels<MarketInternalNavigationViewModel>(R.id.mainFragment)
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -28,16 +25,16 @@ class MarketFragment : BaseWithSearchFragment(), FilterAdapter.Listener {
 
         recyclerTags.adapter = filterAdapter
 
-        filterAdapter.setFilters(marketTabsViewModel.tabs.map { FilterAdapter.FilterItem(it.name) }, FilterAdapter.FilterItem(marketTabsViewModel.currentTab.name))
+        filterAdapter.setFilters(marketViewModel.tabs.map { FilterAdapter.FilterItem(it.name) }, FilterAdapter.FilterItem(marketViewModel.currentTab.name))
 
         viewPager.adapter = MarketTabsAdapter(childFragmentManager, viewLifecycleOwner.lifecycle)
         viewPager.isUserInputEnabled = false
 
-        marketTabsViewModel.tabLiveData.observe(viewLifecycleOwner, { tab: MarketTabsService.Tab ->
+        marketViewModel.tabLiveData.observe(viewLifecycleOwner, { tab: MarketService.Tab ->
             val contentFragment = when (tab) {
-                MarketTabsService.Tab.Overview -> 0
-                MarketTabsService.Tab.Discovery -> 1
-                MarketTabsService.Tab.Favorites -> 2
+                MarketService.Tab.Overview -> 0
+                MarketService.Tab.Discovery -> 1
+                MarketService.Tab.Favorites -> 2
             }
 
             viewPager.setCurrentItem(contentFragment, false)
@@ -46,8 +43,8 @@ class MarketFragment : BaseWithSearchFragment(), FilterAdapter.Listener {
         navigationViewModel.navigateToDiscoveryLiveEvent.observe(viewLifecycleOwner) {
             navigationViewModel.setDiscoveryMode(it)
 
-            marketTabsViewModel.currentTab = MarketTabsService.Tab.Discovery
-            filterAdapter.setFilters(marketTabsViewModel.tabs.map { FilterAdapter.FilterItem(it.name) }, FilterAdapter.FilterItem(marketTabsViewModel.currentTab.name))
+            marketViewModel.currentTab = MarketService.Tab.Discovery
+            filterAdapter.setFilters(marketViewModel.tabs.map { FilterAdapter.FilterItem(it.name) }, FilterAdapter.FilterItem(marketViewModel.currentTab.name))
         }
     }
 
@@ -56,8 +53,8 @@ class MarketFragment : BaseWithSearchFragment(), FilterAdapter.Listener {
     }
 
     override fun onFilterItemClick(item: FilterAdapter.FilterItem?, itemPosition: Int, itemWidth: Int) {
-        MarketTabsService.Tab.fromString(item?.filterId)?.let {
-            marketTabsViewModel.currentTab = it
+        MarketService.Tab.fromString(item?.filterId)?.let {
+            marketViewModel.currentTab = it
         }
     }
 }
