+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/MarketModule.kt
@@ -1,17 +1,17 @@
-package io.horizontalsystems.bankwallet.modules.market.tabs
+package io.horizontalsystems.bankwallet.modules.market
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 
-object MarketTabsModule {
+object MarketModule {
 
     class Factory : ViewModelProvider.Factory {
 
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-            val service = MarketTabsService(App.marketStorage)
-            return MarketTabsViewModel(service) as T
+            val service = MarketService(App.marketStorage)
+            return MarketViewModel(service) as T
         }
 
     }
