+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/MarketViewModel.kt
@@ -1,10 +1,10 @@
-package io.horizontalsystems.bankwallet.modules.market.tabs
+package io.horizontalsystems.bankwallet.modules.market
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 
-class MarketTabsViewModel(private val service: MarketTabsService) : ViewModel() {
+class MarketViewModel(private val service: MarketService) : ViewModel() {
 
     val tabs by service::tabs
     var currentTab by service::currentTab
